refactor(tests): keep provider tests in existing files
